requirements.txt
strands-agents==1.8.0
browser-use
bedrock-agentcore
boto3
python-dotenv
asyncio